Wrapper for ssh
Finds a keyfile associated with an ssh host (using paramiko SSHConfig)
and creates a socket in $XDG_RUNTIME_DIR that is an md5sum of the
filename and only adds that single key. This makes agent forwarding more
secure.

Does make the assumption that the last argument to ssh is the hostname,
other than that it should pass all other arguments to ssh.