Disable automute because pulseaudio
After installing pulseaudio, unsuccessfully trying to hook up a bluetooth
speaker, I rebooted my laptop to find that I had no sound. Turns out
audio was "automuted". This change makes sure that Auto Mute Mode (a mode
the purpose of which is fairly opaque to me) is disabled on xstart.