Static image gallery
I made a thing so that I can upload image galleries. It was inspired by
<https://github.com/brennen/galleryhtml> which was simple and worked
well. This also started as simple and gradually blew up. But I think it
does everything I want.

Basically, I can do:

    publish dir public_html

and it'll look for a `_metadata.yaml` file that has the info about the
page title, images, small paragraph to include, etc. It makes a site and
puts that into public_html.

It also has scripts to generate and upload an index file.

This is what I used to generate: http://photos.tylercipriani.com/