Make notes into tagged blog posts
I've been reading a lot of old blogs. It seems like, in the course of
blogging, folks make pages outside of their blog. These pages tend to
not survive the jumps from platform to platform. The only way to ensure
that you keep all the random crap you've written over the years is to
store it all, journal-like, in a simple log format.