"webify" improvements, new desk photo based on improvements
Sorting through my thoughts on git-photo and how I want to publish my
blog. I think there is an ideal workflow that looks something like:

1. Edit raw files in darktable
2. Export raw to edit
3. This may be a web album to publish on photos.tylercipriani.com or
   this may be private, or this may be junk.
4. git annex add (raw and edit, copy to nas and s3)
5. To publish as a gallery (steps i and ii are interchangeable):
    i. git photo add *.jpg
    ii. git photo init .
    iii. git photo publish .
6. To publish this photo in a blog post:
    i. git photo add [single photo].jpg
    ii. git photo show thumb.large [single photo].jpg
    iii. copy url into post

The problem here is creating a command line tool that relies so heavily
on the config.

Git photo add is uploading stuff to a remote, I guess, else it is just
creating a directory of thumbs for *you* to upload? And then you could
do:

    git photo add thumb.large.url [photo].jpg

I think that makes the most sense, I'll try to get that working.