Standalone image pages
Initially hiraeth was tied to the concept of galleries. Which is fine:
that's generally how I upload files as part of a larger collection of
files.

Then I added the feature of standalone thumbnailing. To share an image
on my blog I do:

    git photo thumb add [file] ~/public_html/thumbs/
    s3cmd sync --acl-public ~/public_html/thumbs/ \
        s3://photos.tylercipriani.com/thumbs/

And that will generate thumbnails that I can use for my site. I can also
retrieve thumbnails via:

    git photo thumb show [file]

Which is neat.

This has led me to realize that photos are a standalone thing. Photos
should be remixable. Added not only to galleries, but tagged lists,
index pages by recency, etc.

If the photo is a standalone, then all of the other pieces are just html
that can be generated. This commit moves towards making each photo a
standalone by giving each photo its own non-gallery page when you add it
as a thumb. This can be used to generate an index page at some point in
the future.