Mail: mutt colors and imapfilter update
Abusing flags as I did previously did not help me track important mail,
rather it just added to the noise. Instead I'm going to rely on flagging
important mail in mutt my using different colors in the index.

To do this, I also need to ensure that the TERM is setup with 256
colors, hence the need to add a new alias.