Literate vimrc file
This is an experiment that was enabled by my giant universal repo of
doom. Like that repo this may, in fact, be a terrible idea. Like that
repo I reserve the right to slowly back away from this idea and mention
that fact to no one except this git log.