blog: add post about git commit messages
This post also contains a long rant about gerrit v github. I have this
discussion quite frequently, and this post should serve as a good point
of reference during those discussions. i.e., it'll be nice to send a
link rather than berate folks with paragraph after paragraph of
Opinions™ (even if that's all the link is).